pyproject.toml
[tool.poetry]
name = "qr-generator"
version = "0.1.0"
description = ""
authors = ["ashshelby <nepalidude3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.7.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
